Add files via upload
I am sorry once again for the delay. I wasn't able to find a function to calculate the diameter and my connection was acting up a lot over the last hour so it took me longer than expected. I am sure one of you found a function for that. I will still be working on my visualization graph to make it more detailed, but this is what I have done so far :)